🎨 (wallet): Restructure for Legibility
